+++ app/models/invoice.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-# Invoice correspond to a single purchase made by an user. This purchase may
-# include reservation(s) and/or a subscription
+# Invoice correspond to a single purchase made by an user. This purchase is linked to one or many invoice_items
 class Invoice < PaymentDocument
   include NotifyWith::NotificationAttachedObject
   require 'fileutils'
@@ -54,7 +53,7 @@ class Invoice < PaymentDocument
 
   # for debug & used by rake task "fablab:maintenance:regenerate_invoices"
   def regenerate_invoice_pdf
-    pdf = ::PDF::Invoice.new(self, subscription&.expiration_date).render
+    pdf = ::PDF::Invoice.new(self, invoice_items.find(&:subscription)&.expiration_date).render
     File.binwrite(file, pdf)
   end
 
@@ -100,7 +99,7 @@ class Invoice < PaymentDocument
 
   def subscription_invoice?
     invoice_items.each do |ii|
-      return true if ii.subscription
+      return true if ii.object_type == 'Subscription'
     end
     false
   end
@@ -130,11 +129,15 @@ class Invoice < PaymentDocument
     return true if user.nil?
 
     # workaround for reservation saved after invoice
-    if invoiced_type == 'Reservation' && invoiced&.reservable_type == 'Training'
-      user.trainings.include?(invoiced.reservable_id)
+    if main_item.object_type == 'Reservation' && main_item.object&.reservable_type == 'Training'
+      user.trainings.include?(main_item.object.reservable_id)
     else
       false
     end
+  end
+
+  def main_item
+    invoice_items.where(main: true).first
   end
 
   # get amount total paid
@@ -168,8 +171,8 @@ class Invoice < PaymentDocument
     return unless Setting.get('invoicing_module')
 
     unless Rails.env.test?
-      puts "Creating an InvoiceWorker job to generate the following invoice: id(#{id}), invoiced_id(#{invoiced_id}), " \
-           "invoiced_type(#{invoiced_type}), user_id(#{invoicing_profile.user_id})"
+      puts "Creating an InvoiceWorker job to generate the following invoice: id(#{id}), main_item.object_id(#{main_item.object_id}), " \
+           "main_item.object_type(#{main_item.object_type}), user_id(#{invoicing_profile.user_id})"
     end
     InvoiceWorker.perform_async(id, user&.subscription&.expired_at)
   end