+++ test/integration/subscriptions/renew_as_admin_test.rb
@@ -68,7 +68,8 @@ class Subscriptions::RenewAsAdminTest < ActionDispatch::IntegrationTest
     assert_equal user.id, notification.receiver_id, 'wrong user notified'
 
     # Check generated invoice
-    invoice = Invoice.find_by(invoiced_type: 'Subscription', invoiced_id: subscription[:id])
+    item = InvoiceItem.find_by(object_type: 'Subscription', object_id: subscription[:id])
+    invoice = item.invoice
     assert_invoice_pdf invoice
     assert_equal plan.amount, invoice.total, 'Invoice total price does not match the bought subscription'
   end